added support for deleting customers

+++ WebContent/WEB-INF/view/list-customers.jsp
@@ -45,7 +45,7 @@
 					<td>${customer.firstName}</td>
 					<td>${customer.lastName}</td>
 					<td>${customer.email}</td>
-					<td><a href="${updateLink}">Update</a> | <a href="${deleteLink}">Delete</a></td>
+					<td><a href="${updateLink}">Update</a> | <a href="${deleteLink}" onclick= "if(!(confirm('Are you sure you want to delete this customer'))) return false;">Delete</a></td>
 				</tr>
 			</c:forEach>
